# spsyncapi/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backuprun.FileTransferDetails:
    properties:
      end_at:
        type: string
      file_path:
        type: string
      start_at:
        type: string
    type: object
  backuprun.RunDetails:
    properties:
      end_at:
        type: string
      id:
        type: string
      job_id:
        type: string
      start_at:
        type: string
    type: object
...
  handlers.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  handlers.backupRunGetResponse:
    properties:
      backup_run:
        $ref: '#/definitions/backuprun.RunDetails'
      file_transfers:
        items:
          $ref: '#/definitions/backuprun.FileTransferDetails'
        type: array
      pagination:
        $ref: '#/definitions/handlers.Pagination'
    type: object
  handlers.backupRunListResponse:
    properties:
      backup_runs:
        items:
          $ref: '#/definitions/backuprun.RunDetails'
        type: array
      pagination:
        $ref: '#/definitions/handlers.Pagination'
    type: object
  handlers.backupRunStartResponse:
    properties:
      backup_run:
        $ref: '#/definitions/backuprun.RunDetails'
    type: object
...
  handlers.createRestoreJobRequest:
    properties:
      active:
        type: boolean
      job_config:
        $ref: '#/definitions/handlers.restoreJobConfigRequest'
      last_run:
        type: string
      start_at:
        type: string
    required:
    - job_config
    type: object
  handlers.errorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  handlers.restoreJobConfigRequest:
    properties:
      bucket_store:
        type: string
      organization:
        type: string
      share_point_site:
        type: string
    required:
    - bucket_store
    - organization
    - share_point_site
    type: object
  handlers.restoreJobListResponse:
    properties:
      restore_jobs:
        items:
          $ref: '#/definitions/restorejob.RestoreJobDetails'
        type: array
    type: object
  handlers.restoreJobResponse:
    properties:
      restore_job:
        $ref: '#/definitions/restorejob.RestoreJobDetails'
    type: object
...
  handlers.successResponse:
    properties:
      success:
        type: boolean
    type: object
...
  handlers.updateRestoreJobRequest:
    properties:
      active:
        type: boolean
      job_config:
        $ref: '#/definitions/handlers.restoreJobConfigRequest'
      last_run:
        type: string
      start_at:
        type: string
    required:
    - job_config
    type: object
...
  restorejob.JobConfigDetails:
    properties:
      bucket_store:
        type: string
      organization:
        type: string
      share_point_site:
        type: string
    type: object
  restorejob.RestoreJobDetails:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      job_config:
        $ref: '#/definitions/restorejob.JobConfigDetails'
      last_run:
        type: string
      start_at:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HTTP API server for the SPSync platform.
  title: SPSync API
  version: "1.0"
paths:
...
  /backup-jobs/{id}/runs:
    post:
      description: Creates a backup run row and starts the Temporal workflow
      parameters:
      - description: Backup job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.backupRunStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Start backup run
      tags:
      - backup-runs
  /backup-runs:
    get:
      description: Returns paginated backup runs; optional job_id filter
      parameters:
      - description: Filter by backup job ID
        in: query
        name: job_id
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.backupRunListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List backup runs
      tags:
      - backup-runs
  /backup-runs/{id}:
    get:
      description: Returns a backup run and paginated file transfers
      parameters:
      - description: Backup run ID
        in: path
        name: id
        required: true
        type: string
      - description: File transfers page (default 1)
        in: query
        name: page
        type: integer
      - description: File transfers page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.backupRunGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get backup run
      tags:
      - backup-runs
  /backup-runs/{id}/stop:
    post:
      description: Cancels the Temporal workflow for an in-progress backup run
      parameters:
      - description: Backup run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.backupRunStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Stop backup run
      tags:
      - backup-runs
...
  /restore-jobs:
    get:
      description: Returns all active restore jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreJobListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List restore jobs
      tags:
      - restore-jobs
    post:
      consumes:
      - application/json
      description: Creates a restore job with job configuration
      parameters:
      - description: Restore job payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createRestoreJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.restoreJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create restore job
      tags:
      - restore-jobs
  /restore-jobs/{id}:
    delete:
      description: Marks a restore job inactive; the record is retained
      parameters:
      - description: Restore job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.successResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete restore job
      tags:
      - restore-jobs
    get:
      description: Returns an active restore job by ID
      parameters:
      - description: Restore job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreJobResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get restore job
      tags:
      - restore-jobs
    put:
      consumes:
      - application/json
      description: Updates a restore job and job config
      parameters:
      - description: Restore job ID
        in: path
        name: id
        required: true
        type: string
      - description: Restore job payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateRestoreJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update restore job
      tags:
      - restore-jobs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"